pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegisap"
version = "0.1.0"
description = "Day 1-5 implementation and Azure training runtime for AegisAP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.135.1",
  "pydantic>=2.7",
  "uvicorn==0.42.0",
  "langsmith==0.7.22",
  "openai==2.29.0",
  "azure-monitor-opentelemetry==1.8.7",
  "azure-identity==1.25.3",
  "azure-keyvault-secrets==4.10.0",
  "azure-search-documents==11.6.0",
  "azure-storage-blob==12.28.0",
  "langgraph==1.1.3",
  "psycopg[binary]==3.3.3",
  "ipykernel>=7.2.0",
  "marimo>=0.9",
  "msal>=1.29",
  "azure-servicebus>=7.12",
  "azure-mgmt-resource>=23.0",
  "azure-mgmt-network>=25.0",
  "httpx>=0.27",
  "tenacity>=8.3",
  "PyJWT>=2.8",
  "requests>=2.32",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
day9 = ["dspy>=2.6"]
day11 = ["msal>=1.29"]
notebooks = ["marimo>=0.9", "altair>=5.0", "pandas>=2.0"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
